pyproject.toml
[project]
name = "automaton-auditor-trp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.46",
    "langchain-xai>=0.0.1",
    "langchain-groq>=0.1.3",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
]
